[BBB-Eurobot] Add driver for color sensor. Change types of some I2C functions.
Most parameters of the I2C drivers should be of type guint8 (i.e. unsigned char i.e. uint8_t) instead of int.
This was changed in I2C-Wrapper and in some other drivers.

+++ ProgrammationBBB/BBBEurobot/include/BBBEurobot/IMUDriver.h
@@ -8,14 +8,14 @@
 #ifndef IMU_DRIVER_H
 	#define IMU_DRIVER_H
 	#include "BBBEurobot/I2C-Wrapper.h"
-	
+
 	/// IMU structure.
 	typedef struct {
 		I2CAdapter *adapter; ///< Pointer to the I2C port being used.
-		int gyroAddress;	 ///< Gyroscope I2C address.
-		int accelAddress;	 ///< Accelerometer I2C address.
+		guint8 gyroAddress;	 ///< Gyroscope I2C address.
+		guint8 accelAddress;	 ///< Accelerometer I2C address.
 	}IMU;
-	
+
 	/// \brief Initialize IMU structure.
     ///
     /// \details This function tests the communication with the IMU, and, if successful, inits the structure.
@@ -26,35 +26,35 @@
     /// \param[in] gyroAddress I2C address of the gyroscope.
     /// \param[in] accelAddress I2C address of the accelerometer.
     /// \returns   TRUE on success, FALSE otherwise.
-	gboolean imu_init(IMU *imu, I2CAdapter *adapter, int gyroAddress, int accelAddress);
-	
+	gboolean imu_init(IMU *imu, I2CAdapter *adapter, guint8 gyroAddress, guint8 accelAddress);
+
 	/// \brief Calls imu_init with the default sensor I2C addresses.
     ///
 	static inline gboolean imu_initDefault(IMU *imu, I2CAdapter *adapter){return imu_init(imu, adapter, 0b1101011, 0b0011101);}
-	
-	
+
+
 	/// \brief Get gyroscope reading along X axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Gyroscope reading along X, in deg/s.
 	double imu_gyroGetXAxis(IMU imu);
-	
-	
+
+
 	/// \brief Get gyroscope reading along Y axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Gyroscope reading along Y, in deg/s.
 	double imu_gyroGetYAxis(IMU imu);
-	
-	
+
+
 	/// \brief Get gyroscope reading along Z axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Gyroscope reading along Z, in deg/s.
 	double imu_gyroGetZAxis(IMU imu);
-	
-	
-	
+
+
+
 	/// \brief Get all gyroscope readings at once.
 	///
     /// \param[in] imu The IMU structure to use for communication.
@@ -62,28 +62,28 @@
     /// \param[out] y Gyroscope reading along Y, in deg/s.
     /// \param[out] z Gyroscope reading along Z, in deg/s.
 	void imu_gyroGetValues(IMU imu, double *x, double *y, double *z);
-	
-	
+
+
 	/// \brief Get accelerometer reading along X axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Accelerometer reading along X, in m/s^2.
 	double imu_accelGetXAxis(IMU imu);
-		
-	
+
+
 	/// \brief Get accelerometer reading along Y axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Accelerometer reading along Y, in m/s^2.
 	double imu_accelGetYAxis(IMU imu);
-		
-	
+
+
 	/// \brief Get accelerometer reading along Z axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Accelerometer reading along Z, in m/s^2.
 	double imu_accelGetZAxis(IMU imu);
-	
+
 
 	/// \brief Get all accelerometer readings at once.
 	///
@@ -92,29 +92,29 @@
     /// \param[out] y Accelerometer reading along Y, in m/s^2.
     /// \param[out] z Accelerometer reading along Z, in m/s^2.
 	void imu_accelGetValues(IMU imu, double *x, double *y, double *z);
-	
-	
+
+
 	/// \brief Get magnetometer reading along X axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Magnetometer reading along X, in mgauss.
 	double imu_magnetoGetXAxis(IMU imu);
-	
-	
+
+
 	/// \brief Get magnetometer reading along Y axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Magnetometer reading along Y, in mgauss.
 	double imu_magnetoGetYAxis(IMU imu);
-	
-	
+
+
 	/// \brief Get magnetometer reading along Z axis.
     ///
     /// \param[in] imu The IMU structure to use for communication.
     /// \returns Magnetometer reading along Z, in mgauss.
 	double imu_magnetoGetZAxis(IMU imu);
-	
-	
+
+
 	/// \brief Get all magnetometer readings at once.
 	///
     /// \param[in] imu The IMU structure to use for communication.
@@ -122,5 +122,5 @@
     /// \param[out] y Magnetometer reading along Y, in mgauss.
     /// \param[out] z Magnetometer reading along Z, in mgauss.
 	void imu_magnetoGetValues(IMU imu, double *x, double *y, double *z);
-	
+
 #endif
